rename verifier files + some cleanups
