+++ src/GraphQL.Analyzers.Tests/Verifiers/CSharpAnalyzerVerifier.Test.cs
@@ -13,8 +13,8 @@ public static partial class CSharpAnalyzerVerifier<TAnalyzer>
         {
             SolutionTransforms.Add((solution, projectId) =>
             {
-                var compilationOptions = solution.GetProject(projectId).CompilationOptions;
-                compilationOptions = compilationOptions.WithSpecificDiagnosticOptions(
+                var compilationOptions = solution.GetProject(projectId)!.CompilationOptions;
+                compilationOptions = compilationOptions!.WithSpecificDiagnosticOptions(
                     compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
                 solution = solution.WithProjectCompilationOptions(projectId, compilationOptions);
