+++ src/GraphQL.Analyzers.Tests/Verifiers/CSharpCodeFixVerifier.Test.cs
@@ -1,20 +1,22 @@
-using Microsoft.CodeAnalysis.CodeRefactorings;
+using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp.Testing;
+using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Testing.Verifiers;
 
 namespace GraphQL.Analyzers.Tests.Verifiers;
 
-public static partial class CSharpCodeRefactoringVerifier<TCodeRefactoring>
-    where TCodeRefactoring : CodeRefactoringProvider, new()
+public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
+    where TAnalyzer : DiagnosticAnalyzer, new()
+    where TCodeFix : CodeFixProvider, new()
 {
-    public class Test : CSharpCodeRefactoringTest<TCodeRefactoring, XUnitVerifier>
+    public class Test : CSharpCodeFixTest<TAnalyzer, TCodeFix, XUnitVerifier>
     {
         public Test()
         {
             SolutionTransforms.Add((solution, projectId) =>
             {
-                var compilationOptions = solution.GetProject(projectId).CompilationOptions;
-                compilationOptions = compilationOptions.WithSpecificDiagnosticOptions(
+                var compilationOptions = solution.GetProject(projectId)!.CompilationOptions;
+                compilationOptions = compilationOptions!.WithSpecificDiagnosticOptions(
                     compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
                 solution = solution.WithProjectCompilationOptions(projectId, compilationOptions);
