+++ src/GraphQL.Analyzers.Tests/Verifiers/CSharpCodeRefactoringVerifier.cs
